refactor subscription provider logic
